The Two Component Mixture Normal Distribution

The Two Component Mixture Normal Distribution

Usage

pmnorm(q, mu, sigma, lambda, log = FALSE, verbose = FALSE)

dmnorm(x, mu, sigma, lambda, log = FALSE, verbose = FALSE)

Arguments

q, x

(Vector). Values to calculate distributional values of.

mu

(Vector). A two value vector of mean values.

sigma

(Vector). A two value vector of component-wise variances

lambda

(Vector). A two value vector of component mixing coefficients

log

(Boolean). Whether returning probabilities are in log format

verbose

(Boolean). Whether to return component values.

Value

A numeric value representing the probability density value of a two-component mixture distribution

Functions

pmnorm: Cumulative Distribution Function

dmnorm: Probability Density Function

Examples

    library(mixtools)
#> mixtools package, version 1.2.0, Released 2020-02-05
#> This package is based upon work supported by the National Science Foundation under Grant No. SES-0518772.
    lambda <- c(0.7,0.3)
    mu <- c(1,2)
    sigma <- c(1,1)
    v <- rnormmix(100, lambda=lambda, mu=mu, sigma=sigma)
    pmnorm(v, lambda=lambda,mu=mu,sigma=sigma)
#>   [1] 0.157682532 0.209710558 0.945249302 0.468998916 0.417523336 0.198063829
#>   [7] 0.852510710 0.212313883 0.845538768 0.478135296 0.027126524 0.691400538
#>  [13] 0.257568727 0.899553684 0.225759967 0.062040716 0.368891479 0.035159292
#>  [19] 0.821859101 0.861335834 0.326543187 0.720961260 0.690504512 0.523966000
#>  [25] 0.362798836 0.422291544 0.772437882 0.138359027 0.489722853 0.783458213
#>  [31] 0.247592197 0.657705102 0.011160023 0.946029507 0.249710996 0.552832362
#>  [37] 0.382609474 0.506715254 0.847148032 0.248825746 0.579937807 0.177880167
#>  [43] 0.113497141 0.418839304 0.856362123 0.160752741 0.737496487 0.909557943
#>  [49] 0.254891703 0.489444084 0.155868611 0.787247581 0.008454851 0.562379710
#>  [55] 0.146753925 0.041394748 0.499962949 0.788985273 0.417110908 0.953368355
#>  [61] 0.815265137 0.105539348 0.385031858 0.688025692 0.808203482 0.735296639
#>  [67] 0.810269366 0.080860460 0.389643391 0.008117310 0.704811263 0.534487987
#>  [73] 0.739059527 0.896620723 0.044327822 0.956506694 0.904286138 0.467725408
#>  [79] 0.722831659 0.746356757 0.893253052 0.679658708 0.460595344 0.232406640
#>  [85] 0.751213678 0.578190037 0.886211425 0.391619228 0.940613132 0.207339864
#>  [91] 0.475447637 0.897300889 0.579826424 0.774661203 0.999068857 0.996921870
#>  [97] 0.684946744 0.797834826 0.926436558 0.375124342
    dmnorm(v, lambda=lambda,mu=mu,sigma=sigma)
#>   [1] 0.225203026 0.267468335 0.098329903 0.360635731 0.355492823 0.258847966
#>   [7] 0.204008023 0.269332868 0.210412243 0.360916542 0.059853298 0.314035438
#>  [13] 0.298305603 0.155768889 0.278610470 0.115945075 0.344940294 0.074005056
#>  [19] 0.230898358 0.195644167 0.331040508 0.299178766 0.314452088 0.359492285
#>  [25] 0.343216193 0.356226142 0.267931751 0.206840605 0.361001721 0.260301495
#>  [31] 0.292458988 0.328374846 0.028241427 0.097229097 0.293725463 0.356183987
#>  [37] 0.348490607 0.360580894 0.208949606 0.293197956 0.351382146 0.242788236
#>  [43] 0.180759200 0.355700511 0.200393881 0.227978592 0.289906078 0.144226046
#>  [49] 0.296765736 0.361003221 0.223545379 0.257597227 0.022197873 0.354680242
#>  [55] 0.215011516 0.084404674 0.360825662 0.256343140 0.355426900 0.086626327
#>  [61] 0.236271679 0.171766918 0.349070209 0.315594555 0.241873078 0.291180473
#>  [67] 0.240250560 0.141613229 0.350134619 0.021422850 0.307564466 0.358501849
#>  [73] 0.288992909 0.159058774 0.089141214 0.081946120 0.150371775 0.360581587
#>  [79] 0.298165077 0.284644655 0.162785777 0.319340970 0.360210472 0.282982981
#>  [85] 0.281672215 0.351741425 0.170409762 0.350575075 0.104773303 0.265750793
#>  [91] 0.360853528 0.158299518 0.351405247 0.266420158 0.002904877 0.008535395
#>  [97] 0.316992858 0.249818168 0.123525172 0.346610075