The Two Component Mixture Normal Distribution

The Two Component Mixture Normal Distribution

Usage

pmnorm(q, mu, sigma, lambda, log = FALSE, verbose = FALSE)

dmnorm(x, mu, sigma, lambda, log = FALSE, verbose = FALSE)

Arguments

q, x

(Vector). Values to calculate distributional values of.

mu

(Vector). A two value vector of mean values.

sigma

(Vector). A two value vector of component-wise variances

lambda

(Vector). A two value vector of component mixing coefficients

log

(Boolean). Whether returning probabilities are in log format

verbose

(Boolean). Whether to return component values.

Value

A numeric value representing the probability density value of a two-component mixture distribution

Functions

pmnorm: Cumulative Distribution Function

dmnorm: Probability Density Function

Examples

    library(mixtools)
#> mixtools package, version 1.2.0, Released 2020-02-05
#> This package is based upon work supported by the National Science Foundation under Grant No. SES-0518772.
    lambda <- c(0.7,0.3)
    mu <- c(1,2)
    sigma <- c(1,1)
    v <- rnormmix(100, lambda=lambda, mu=mu, sigma=sigma)
    pmnorm(v, lambda=lambda,mu=mu,sigma=sigma)
#>   [1] 0.40760891 0.31100758 0.80884167 0.91974830 0.63744738 0.41419654
#>   [7] 0.71301969 0.23803560 0.21690062 0.60931000 0.05610485 0.52919234
#>  [13] 0.93662715 0.37421253 0.35087402 0.84785113 0.55226348 0.24131129
#>  [19] 0.32503969 0.43829462 0.89924360 0.66736489 0.42875302 0.82650011
#>  [25] 0.64781734 0.26802695 0.17727384 0.03291118 0.15559892 0.45142645
#>  [31] 0.57192482 0.32385631 0.29685083 0.38747681 0.01825383 0.43729567
#>  [37] 0.84531777 0.30437828 0.69517717 0.23319644 0.96394350 0.85674974
#>  [43] 0.79748104 0.20198046 0.90246200 0.48654972 0.65872784 0.86056166
#>  [49] 0.99056954 0.01042646 0.65380867 0.35324187 0.89309279 0.37444903
#>  [55] 0.08819896 0.47244133 0.22529989 0.81940714 0.15808793 0.15238785
#>  [61] 0.27615879 0.12047299 0.17909570 0.87025299 0.46362245 0.04714138
#>  [67] 0.84208187 0.78058510 0.72006477 0.61704104 0.71415137 0.37479715
#>  [73] 0.03685410 0.95441179 0.93587624 0.54248010 0.02675772 0.93064593
#>  [79] 0.25357111 0.54847714 0.53188175 0.51072567 0.10376338 0.62444043
#>  [85] 0.15112258 0.22226156 0.51274217 0.57337936 0.56701556 0.98841579
#>  [91] 0.46521512 0.83058171 0.16004026 0.13887958 0.33523459 0.29772974
#>  [97] 0.59142922 0.61486546 0.72788287 0.37438433
    dmnorm(v, lambda=lambda,mu=mu,sigma=sigma)
#>   [1] 0.35379771 0.32479264 0.24137327 0.13192386 0.33570499 0.35494973
#>   [7] 0.30338414 0.28657803 0.27256359 0.34431453 0.10728050 0.35903123
#>  [13] 0.11018505 0.34637171 0.33957632 0.20830766 0.35626716 0.28862515
#>  [19] 0.33046337 0.35830130 0.15611868 0.32454086 0.35713542 0.22703197
#>  [25] 0.33207216 0.30412080 0.24228296 0.07013606 0.22329779 0.35956322
#>  [31] 0.35297300 0.33000499 0.31854279 0.34964090 0.04299187 0.35818909
#>  [37] 0.21061238 0.32193284 0.31225779 0.28349334 0.07046394 0.20002708
#>  [43] 0.25008350 0.26179829 0.15246528 0.36100846 0.32797934 0.19638966
#>  [49] 0.02282932 0.02663003 0.32985916 0.34032720 0.16296181 0.34643373
#>  [55] 0.15096211 0.36076370 0.27830264 0.23291275 0.22557168 0.22032699
#>  [61] 0.30842511 0.18837552 0.24379706 0.18688903 0.36038213 0.09359846
#>  [67] 0.21352302 0.26232420 0.29966148 0.34212942 0.30279458 0.34652477
#>  [73] 0.07687849 0.08508045 0.11119186 0.35758438 0.05918045 0.11809769
#>  [79] 0.29599832 0.35680230 0.35877020 0.36038719 0.16971420 0.33991032
#>  [85] 0.21914476 0.27625280 0.36027619 0.35269493 0.35387655 0.02725481
#>  [91] 0.36046409 0.22357284 0.22733784 0.20735619 0.33426267 0.31894650
#>  [97] 0.34884890 0.34275809 0.29538288 0.34641678