The Two Component Mixture Normal Distribution

The Two Component Mixture Normal Distribution

Usage

pmnorm(q, mu, sigma, lambda, log = FALSE, verbose = FALSE)

dmnorm(x, mu, sigma, lambda, log = FALSE, verbose = FALSE)

Arguments

q, x

(Vector). Values to calculate distributional values of.

mu

(Vector). A two value vector of mean values.

sigma

(Vector). A two value vector of component-wise variances

lambda

(Vector). A two value vector of component mixing coefficients

log

(Boolean). Whether returning probabilities are in log format

verbose

(Boolean). Whether to return component values.

Value

A numeric value representing the probability density value of a two-component mixture distribution

Functions

pmnorm: Cumulative Distribution Function

dmnorm: Probability Density Function

Examples

    library(mixtools)
#> mixtools package, version 1.2.0, Released 2020-02-05
#> This package is based upon work supported by the National Science Foundation under Grant No. SES-0518772.
    lambda <- c(0.7,0.3)
    mu <- c(1,2)
    sigma <- c(1,1)
    v <- rnormmix(100, lambda=lambda, mu=mu, sigma=sigma)
    pmnorm(v, lambda=lambda,mu=mu,sigma=sigma)
#>   [1] 0.040283172 0.617588015 0.844238862 0.240060528 0.866550483 0.475537916
#>   [7] 0.592262648 0.637901613 0.576691723 0.263201065 0.179883787 0.817324211
#>  [13] 0.233780360 0.185948786 0.544541217 0.384973486 0.975426595 0.079534482
#>  [19] 0.535294375 0.637402413 0.106592660 0.644712904 0.516370486 0.628821674
#>  [25] 0.507904875 0.082181491 0.944859712 0.196100931 0.675274523 0.552383481
#>  [31] 0.297903767 0.958631438 0.696145301 0.243751127 0.074908097 0.684183249
#>  [37] 0.776607800 0.113123935 0.415887760 0.876155051 0.158945270 0.364544101
#>  [43] 0.894310104 0.403472994 0.885610976 0.299646354 0.048909205 0.587306761
#>  [49] 0.537983823 0.812177779 0.838339027 0.643719810 0.349735103 0.413893783
#>  [55] 0.422937989 0.360722689 0.058457636 0.243581918 0.831095292 0.903829547
#>  [61] 0.492467103 0.597014722 0.913958971 0.042501047 0.375413276 0.469418323
#>  [67] 0.806738876 0.643737456 0.706105456 0.177796426 0.392757861 0.879279568
#>  [73] 0.260889704 0.362982065 0.509806235 0.826937189 0.727205286 0.322538081
#>  [79] 0.742164806 0.586486072 0.545914316 0.561726872 0.898064292 0.599126218
#>  [85] 0.653839888 0.535747474 0.004482895 0.172783583 0.515174085 0.162121466
#>  [91] 0.526758280 0.426610614 0.507348896 0.059681786 0.075450180 0.672878000
#>  [97] 0.018367448 0.602543466 0.499682965 0.532966835
    dmnorm(v, lambda=lambda,mu=mu,sigma=sigma)
#>   [1] 0.08258456 0.34196966 0.21158696 0.28784737 0.19056227 0.36085591
#>   [7] 0.34865361 0.33555108 0.35204397 0.30147685 0.24444822 0.23460874
#>  [13] 0.28386940 0.24938343 0.35732465 0.34905641 0.05144581 0.13988680
#>  [19] 0.35841575 0.33572020 0.17297633 0.33318590 0.36005358 0.33853727
#>  [25] 0.36052719 0.14332168 0.09887777 0.25734946 0.32123676 0.35624967
#>  [31] 0.31902619 0.07872383 0.31179648 0.29012861 0.13377021 0.31733605
#>  [37] 0.26508524 0.18034483 0.35522905 0.18091901 0.22634915 0.34371939
#>  [43] 0.16162167 0.35302223 0.17104955 0.31981962 0.09635773 0.34979186
#>  [49] 0.35811808 0.23873988 0.21684379 0.33353745 0.33921016 0.35489901
#>  [55] 0.35632148 0.34260782 0.11075058 0.29002492 0.22313365 0.15089736
#>  [61] 0.36097764 0.34751014 0.13898326 0.08620231 0.34668519 0.36065276
#>  [67] 0.24301533 0.33353122 0.30691652 0.24271853 0.35082466 0.17770010
#>  [73] 0.30018665 0.34326937 0.36043478 0.22666419 0.29575988 0.32949010
#>  [79] 0.28715971 0.34997502 0.35714635 0.35478956 0.15744471 0.34698570
#>  [85] 0.32984741 0.35836673 0.01269868 0.23849839 0.36013024 0.22920386
#>  [91] 0.35925354 0.35684468 0.36055268 0.11253731 0.13449452 0.32225360
#>  [97] 0.04321767 0.34611557 0.36083361 0.35866032