The Two Component Mixture Normal Distribution

The Two Component Mixture Normal Distribution

Usage

pmnorm(q, mu, sigma, lambda, log = FALSE, verbose = FALSE)

dmnorm(x, mu, sigma, lambda, log = FALSE, verbose = FALSE)

Arguments

q, x

(Vector). Values to calculate distributional values of.

mu

(Vector). A two value vector of mean values.

sigma

(Vector). A two value vector of component-wise variances

lambda

(Vector). A two value vector of component mixing coefficients

log

(Boolean). Whether returning probabilities are in log format

verbose

(Boolean). Whether to return component values.

Value

A numeric value representing the probability density value of a two-component mixture distribution

Functions

pmnorm: Cumulative Distribution Function

dmnorm: Probability Density Function

Examples

    library(mixtools)
#> mixtools package, version 1.2.0, Released 2020-02-05
#> This package is based upon work supported by the National Science Foundation under Grant No. SES-0518772.
    lambda <- c(0.7,0.3)
    mu <- c(1,2)
    sigma <- c(1,1)
    v <- rnormmix(100, lambda=lambda, mu=mu, sigma=sigma)
    pmnorm(v, lambda=lambda,mu=mu,sigma=sigma)
#>   [1] 0.00276636 0.31769645 0.60981041 0.65793591 0.66902711 0.60798673
#>   [7] 0.02579984 0.66516971 0.58278043 0.08017439 0.69025848 0.50295528
#>  [13] 0.82996433 0.60390548 0.38186572 0.37608928 0.66172597 0.35821744
#>  [19] 0.23180662 0.29438160 0.77928094 0.44665096 0.69014820 0.49676818
#>  [25] 0.57419007 0.25202544 0.16363544 0.17474506 0.47148225 0.86131961
#>  [31] 0.17215549 0.73310211 0.52082471 0.12638829 0.61469616 0.92302081
#>  [37] 0.46645495 0.28440048 0.57051661 0.51646408 0.17338920 0.93749647
#>  [43] 0.54161833 0.93188402 0.37729798 0.15420907 0.40434908 0.54222369
#>  [49] 0.69441360 0.98443942 0.43033350 0.07035825 0.74841399 0.37758362
#>  [55] 0.23982026 0.87886037 0.92253624 0.17848040 0.95470608 0.52420452
#>  [61] 0.21084422 0.27830493 0.17705501 0.95344190 0.14354188 0.95293916
#>  [67] 0.53467633 0.77024401 0.56175531 0.46946377 0.23630925 0.38000003
#>  [73] 0.45569920 0.69066329 0.61769075 0.78341351 0.46537432 0.46302506
#>  [79] 0.02469224 0.53522415 0.27728051 0.53365485 0.84943788 0.67065454
#>  [85] 0.89637157 0.93388837 0.65483173 0.11169881 0.77157244 0.05366823
#>  [91] 0.59631026 0.32705568 0.63108605 0.62108620 0.54428477 0.58729678
#>  [97] 0.94116311 0.25537891 0.10588432 0.89395002
    dmnorm(v, lambda=lambda,mu=mu,sigma=sigma)
#>   [1] 0.00825061 0.32756013 0.34417722 0.32828580 0.32385886 0.34467492
#>   [7] 0.05742251 0.32543149 0.35078315 0.14072142 0.31456615 0.36072980
#>  [13] 0.22409963 0.34576139 0.34830981 0.34686012 0.32680572 0.34185952
#>  [19] 0.28259396 0.31739736 0.26323452 0.35915007 0.31461723 0.36090585
#>  [25] 0.35253788 0.29509354 0.23055054 0.24016091 0.36073074 0.19565981
#>  [31] 0.23796296 0.29243924 0.35974005 0.19464738 0.34280656 0.12784802
#>  [37] 0.36052391 0.31259712 0.35323771 0.36004745 0.23901325 0.10901451
#>  [43] 0.35769017 0.11647955 0.34717015 0.22201711 0.35318986 0.35761603
#>  [49] 0.31262001 0.03505914 0.35734308 0.12760863 0.28339331 0.34724290
#>  [55] 0.28769742 0.17813435 0.12845555 0.24328708 0.08464264 0.35947257
#>  [61] 0.26828304 0.30952984 0.24210028 0.08651770 0.21192110 0.08725930
#>  [67] 0.35848187 0.26940970 0.35478482 0.36065458 0.28548589 0.34785042
#>  [73] 0.35988869 0.31437840 0.34193958 0.26033315 0.36047197 0.36034991
#>  [79] 0.05537064 0.35842330 0.30900413 0.35858928 0.20685235 0.32318476
#>  [85] 0.15933635 0.11383833 0.32947288 0.17875596 0.26851639 0.10363536
#>  [91] 0.34768287 0.33123590 0.33781035 0.34093178 0.35735748 0.34979410
#>  [97] 0.10401749 0.29704756 0.17216367 0.16201881